feat(rust): consolidate RIIR backend implementation

+++ komga-rust/crates/interfaces/src/opds/types.rs
@@ -16,7 +16,7 @@ pub(super) struct PersistedSeriesSearchResult {
     pub(super) id: String,
     pub(super) title: String,
     pub(super) library_id: String,
-    pub(super) age_rating: Option<u16>,
+    pub(super) age_rating: Option<u32>,
     pub(super) sharing_labels: Vec<String>,
     pub(super) last_modified: String,
 }
